Weather API displayed in Header per design
