--- client/src/components/Weather/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Weather";
